feat(submitentityclaim): finish location selector
